refactor: split shared text utils and prompts out of executor_lightbar.go
